IPAAnsibleModule: set _ansible_debug in constructor.
By setting _ansible_debug in the constructor of IPAAnsibleModule class
to a default value, all modules will use the same value, with no effort.

As the value for the object was being set after object constuctor, if
the value should have a different value, it still can be set exactly as
it was before.